move aws to conmmon
