+++ common/app/fronts/FrontsApi.scala
@@ -3,13 +3,16 @@ package fronts
 import com.amazonaws.services.s3.AmazonS3Client
 import com.gu.facia.client.{AmazonSdkS3Client, ApiClient}
 import common.ExecutionContexts
-import conf.Configuration
+import conf.{Configuration, aws}
 import services.AwsEndpoints
 
 object FrontsApi extends ExecutionContexts {
   val amazonClient: ApiClient = {
-    val client = new AmazonS3Client(Configuration.aws.credentials.get)
+    val client = new AmazonS3Client(aws.mandatoryCredentials)
     client.setEndpoint(AwsEndpoints.s3)
-    ApiClient("aws-frontend-store", Configuration.facia.stage.toUpperCase, AmazonSdkS3Client(client))
+    val bucket = Configuration.aws.bucket;
+    val stage = Configuration.facia.stage.toUpperCase;
+    println("* S3 config on bucket " + bucket + " and stage " + stage)
+    ApiClient(bucket, stage, AmazonSdkS3Client(client))
   }
 }
